chore(panel): rename UserDashboardDetails to DetailsPanel
