+++ src/containers/DetailsPanel/DetailsPanelHeader/DetailsPanelHeader.styled.js
