+++ src/containers/DetailsPanel/DetailsPanelSlideOut/DetailsPanelSlideOut.styled.js
