+++ src/pages/UserDashboardPage/UserDashboardTasks/UserTasksContainer.jsx
@@ -8,11 +8,11 @@ import UserDashboardKanBan from './UserDashboardKanBan'
 import { useEffect, useMemo } from 'react'
 import { onAssigneesChanged } from '/src/features/dashboard'
 import { Splitter, SplitterPanel } from 'primereact/splitter'
-import UserDashboardDetails from './UserDashboardDetails/UserDashboardDetails'
+import DetailsPanel from '../../../containers/DetailsPanel/DetailsPanel'
 import { getIntersectionFields, getMergedFields } from '../util'
 import { Section } from '@ynput/ayon-react-components'
 import { setUri } from '/src/features/context'
-import UserDashboardSlideOut from './UserDashboardSlideOut/UserDashboardSlideOut'
+import DetailsPanelSlideOut from '../../../containers/DetailsPanel/DetailsPanelSlideOut/DetailsPanelSlideOut'
 
 export const getThumbnailUrl = ({ entityId, entityType, thumbnailId, updatedAt, projectName }) => {
   if (!projectName || (!thumbnailId && !entityId)) return null
@@ -224,7 +224,7 @@ const UserTasksContainer = ({ projectsInfo = {}, isLoadingInfo }) => {
             minWidth: detailsMinWidth,
           }}
         >
-          <UserDashboardDetails
+          <DetailsPanel
             entitiesData={selectedTasksData}
             statusesOptions={statusesOptions}
             disabledStatuses={disabledStatuses}
@@ -238,7 +238,7 @@ const UserTasksContainer = ({ projectsInfo = {}, isLoadingInfo }) => {
             entityType="task"
             style={{ zIndex: 400 }}
           />
-          <UserDashboardSlideOut projectsInfo={projectsInfo} />
+          <DetailsPanelSlideOut projectsInfo={projectsInfo} />
         </SplitterPanel>
       ) : (
         <SplitterPanel style={{ maxWidth: 0 }}></SplitterPanel>
